Implement support for the SO_LINGER sockopt.
This requires DEFINE-SOCKET-OPTION to be modified slightly, so that arg-
mangling functions are also passed the alien buffer; functions that wish
to modify the existing buffer may do so any return it, while simple
value-converting functions can just return a new value.

This includes two tests for the new option: a simple one verifying that
values can be set and retrieved for a socket, and a more complex test
verifying that setting a linger time of 0 causes a connection to be
hard-closed, flushing any unread data in the process.